requirements.txt
Django==1.9.7
Pygments==2.1.3
djangorestframework==3.3.3
httpie==0.9.4

#You can use this one, but if you running project on python 3.4 you should better use mysqlclient
#MySQL-python==1.2.5

mysqlclient==1.3.7
pip==7.1.0
requests==2.10.0
setuptools==18.0.1
wheel==0.24.0